split into rbac option file and add option combination logic
