+++ api/types/access_policy.go
@@ -17,7 +17,6 @@ limitations under the License.
 package types
 
 import (
-	"strconv"
 	"time"
 
 	"github.com/gravitational/teleport/api/defaults"
@@ -34,21 +33,11 @@ type AccessPolicy interface {
 	GetDeny() map[string]string
 }
 
-// AsOption describes a value that can be deserialized from a predicate option raw value.
-type AsOption interface {
-	// Name is the static name of the option type.
-	Name() string
-
-	// DeserializeInto deserializes the raw value into the receiver.
-	DeserializeInto(raw string) error
-}
-
 // AccessPolicyOption retrieves and attempts to deserialize it into the provided type.
-func AccessPolicyOption[T AsOption](policy *AccessPolicyV1) (T, error) {
+func AccessPolicyOption[T FromRawOption](policy *AccessPolicyV1) (T, error) {
 	var opt T
 	raw, ok := policy.Spec.Options[opt.Name()]
 	if !ok {
-		// TODO: default here?
 		return opt, trace.NotFound("option %q not found", opt.Name())
 	}
 
@@ -57,283 +46,6 @@ func AccessPolicyOption[T AsOption](policy *AccessPolicyV1) (T, error) {
 	}
 
 	return opt, nil
-}
-
-func deserializeOptionDuration(raw string) (time.Duration, error) {
-	i, err := strconv.ParseInt(raw, 10, 64)
-	if err != nil {
-		return 0, trace.BadParameter("invalid duration %q", raw)
-	}
-
-	return time.Duration(i), nil
-}
-
-func deserializeOptionBool(raw string) (bool, error) {
-	b, err := strconv.ParseBool(raw)
-	if err != nil {
-		return false, trace.BadParameter("invalid duration %q", raw)
-	}
-
-	return b, nil
-}
-
-func deserializeOptionInt(raw string) (int, error) {
-	b, err := strconv.Atoi(raw)
-	if err != nil {
-		return 0, trace.BadParameter("invalid integer %q", raw)
-	}
-
-	return b, nil
-}
-
-type AccessPolicySessionTTL time.Duration
-
-// Name is the static name of the option type.
-func (o *AccessPolicySessionTTL) Name() string {
-	return "session_ttl"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySessionTTL) DeserializeInto(raw string) error {
-	d, err := deserializeOptionDuration(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySessionTTL(d)
-	return nil
-}
-
-type AccessPolicyLockingMode int
-
-const (
-	AccessPolicyLockingModeBestEffort AccessPolicyLockingMode = iota
-	AccessPolicyLockingModeStrict
-)
-
-// Name is the static name of the option type.
-func (o *AccessPolicyLockingMode) Name() string {
-	return "locking_mode"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicyLockingMode) DeserializeInto(raw string) error {
-	switch raw {
-	case "best_effort":
-		*o = AccessPolicyLockingModeBestEffort
-		return nil
-	case "strict":
-		*o = AccessPolicyLockingModeStrict
-		return nil
-	default:
-		return trace.BadParameter("invalid locking mode %q", raw)
-	}
-}
-
-type AccessPolicySessionMFA bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySessionMFA) Name() string {
-	return "session_mfa"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySessionMFA) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySessionMFA(b)
-	return nil
-}
-
-type AccessPolicySSHSessionRecordingMode int
-
-const (
-	AccessPolicySSHSessionRecordingModeBestEffort AccessPolicySSHSessionRecordingMode = iota
-	AccessPolicySSHSessionRecordingModeStrict
-)
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHSessionRecordingMode) Name() string {
-	return "ssh.session_recording_mode"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHSessionRecordingMode) DeserializeInto(raw string) error {
-	switch raw {
-	case "best_effort":
-		*o = AccessPolicySSHSessionRecordingModeBestEffort
-		return nil
-	case "strict":
-		*o = AccessPolicySSHSessionRecordingModeStrict
-		return nil
-	default:
-		return trace.BadParameter("invalid session recording mode %q", raw)
-	}
-}
-
-type AccessPolicySSHAllowAgentForwarding bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHAllowAgentForwarding) Name() string {
-	return "ssh.allow_agent_forwarding"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHAllowAgentForwarding) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHAllowAgentForwarding(b)
-	return nil
-}
-
-type AccessPolicySSHAllowPortForwarding bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHAllowPortForwarding) Name() string {
-	return "ssh.allow_port_forwarding"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHAllowPortForwarding) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHAllowPortForwarding(b)
-	return nil
-}
-
-type AccessPolicySSHAllowX11Forwarding bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHAllowX11Forwarding) Name() string {
-	return "ssh.allow_x11_forwarding"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHAllowX11Forwarding) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHAllowX11Forwarding(b)
-	return nil
-}
-
-type AccessPolicySSHAllowFileCopying bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHAllowFileCopying) Name() string {
-	return "ssh.allow_file_copying"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHAllowFileCopying) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHAllowFileCopying(b)
-	return nil
-}
-
-type AccessPolicySSHAllowExpiredCert bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHAllowExpiredCert) Name() string {
-	return "ssh.allow_expired_cert"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHAllowExpiredCert) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHAllowExpiredCert(b)
-	return nil
-}
-
-type AccessPolicySSHPinSourceIP bool
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHPinSourceIP) Name() string {
-	return "ssh.pin_source_ip"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHPinSourceIP) DeserializeInto(raw string) error {
-	b, err := deserializeOptionBool(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHPinSourceIP(b)
-	return nil
-}
-
-type AccessPolicySSHMaxConnections int
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHMaxConnections) Name() string {
-	return "ssh.max_connections"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHMaxConnections) DeserializeInto(raw string) error {
-	i, err := deserializeOptionInt(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHMaxConnections(i)
-	return nil
-}
-
-type AccessPolicySSHMaxSessionsPerConnection int
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHMaxSessionsPerConnection) Name() string {
-	return "ssh.max_sessions_per_connection"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHMaxSessionsPerConnection) DeserializeInto(raw string) error {
-	i, err := deserializeOptionInt(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHMaxSessionsPerConnection(i)
-	return nil
-}
-
-type AccessPolicySSHClientIdleTimeout time.Duration
-
-// Name is the static name of the option type.
-func (o *AccessPolicySSHClientIdleTimeout) Name() string {
-	return "ssh.client_idle_timeout"
-}
-
-// DeserializeInto deserializes the raw value into the receiver.
-func (o *AccessPolicySSHClientIdleTimeout) DeserializeInto(raw string) error {
-	d, err := deserializeOptionDuration(raw)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	*o = AccessPolicySSHClientIdleTimeout(d)
-	return nil
 }
 
 // NewAccessPolicy creates a new access policy from a specification.
